% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_gnod.R
\name{clean_url_to_id}
\alias{clean_url_to_id}
\title{Clean URL to ID}
\usage{
clean_url_to_id(urls)
}
\arguments{
\item{urls}{A character vector of URLs to be cleaned.}
}
\value{
A character vector of cleaned IDs.
}
\description{
This function takes a vector of URLs as strings, cleans them by replacing specific characters with "_",
decodes percentages and some specific characters, and removes special characters by replacing them
with nothing to form clean identifiers. It is designed to convert URL slugs or query parameters into
more readable or database-friendly IDs.
}
\examples{
urls <- c("the+beatles", "queen", "pink+floyd", "simon+\%26+garfunkel")
clean_url_to_id(urls)
}
