% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_census_data.R
\name{get_datagotchi_usa_population_data}
\alias{get_datagotchi_usa_population_data}
\title{Retrieve Population Data by Geographical Division in the USA}
\usage{
get_datagotchi_usa_population_data(geo_division)
}
\arguments{
\item{geo_division}{A character string specifying the geographical division for which data is to be retrieved, such as states or counties.}
}
\value{
A dataframe containing the geographical identifier (GEOID), name of the geographical area (NAME), variable names,
        categories split from those variables, and the proportion of each category within the specified geographical division.
}
\description{
This function collects and processes U.S. population data based on geographical divisions.
It acquires data across various demographic, socio-economic, and education variables from the American Community Survey (ACS).
The output is a dataframe with geographical identifiers and proportionate distributions of the population across specified categories.
}
\examples{
# Assuming 'get_acs' and all transformations are defined and functional within your environment:
df_population = get_datagotchi_usa_population_data("california")

}
