% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_gnod.R
\name{extract_gnod_closeness_df}
\alias{extract_gnod_closeness_df}
\title{Fetch Closeness Matrix}
\usage{
extract_gnod_closeness_df(page, item_url_id)
}
\arguments{
\item{url_id}{A string representing the base URL of the item's page to scrape from.}
}
\value{
A square matrix where each row and column represents an item (artist, book, movie, or series),
        and the cell values represent the closeness scores between each pair of items.
}
\description{
Retrieves and constructs a matrix representing the "closeness" of related items (such as artists, books, movies, or series)
based on a specified base URL. It scrapes the names, URLs, and their closeness scores from the HTML content of the page.
}
\examples{
url_id <- "http://some-base-url.com/item-name"
closeness_matrix <- fetch_closeness_matrix(url_id)
}
